extraction event indexer
