+++ infrastructure/src/main/java/org/hesperides/infrastructure/redis/eventstores/LegacyRedisStorageEngine.java
@@ -1,6 +1,5 @@
 package org.hesperides.infrastructure.redis.eventstores;
 
-import com.google.common.collect.Lists;
 import lombok.extern.slf4j.Slf4j;
 import org.axonframework.common.Assert;
 import org.axonframework.eventhandling.EventMessage;
@@ -11,26 +10,19 @@ import org.axonframework.eventsourcing.eventstore.TrackedEventData;
 import org.axonframework.eventsourcing.eventstore.TrackingToken;
 import org.axonframework.serialization.Serializer;
 import org.hesperides.domain.security.UserEvent;
-import org.springframework.core.io.ClassPathResource;
 import org.springframework.dao.DataAccessException;
 import org.springframework.data.redis.connection.RedisZSetCommands;
 import org.springframework.data.redis.core.RedisOperations;
 import org.springframework.data.redis.core.SessionCallback;
 import org.springframework.data.redis.core.StringRedisTemplate;
-import org.springframework.data.redis.core.script.DefaultRedisScript;
-import org.springframework.scripting.support.ResourceScriptSource;
 import org.springframework.stereotype.Component;
-import org.springframework.util.StopWatch;
 
 import javax.annotation.PostConstruct;
 import javax.annotation.PreDestroy;
 import java.util.*;
-import java.util.concurrent.*;
-import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Consumer;
 import java.util.function.Function;
 import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 import java.util.stream.Stream;
 import java.util.stream.StreamSupport;
 
@@ -70,7 +62,7 @@ public class LegacyRedisStorageEngine extends AbstractEventStorageEngine {
 
     private final StringRedisTemplate template;
     private final Codec codec;
-    private final EventsIndex eventsIndex;
+    private final EventsIndexer eventsIndexer;
     private final EventsListener eventsListener;
 
     public LegacyRedisStorageEngine(StringRedisTemplate template, Codec codec) {
@@ -78,7 +70,7 @@ public class LegacyRedisStorageEngine extends AbstractEventStorageEngine {
         this.template = template;
         this.codec = codec;
 
-        this.eventsIndex = new EventsIndex();
+        this.eventsIndexer = new EventsIndexer(template);
         this.eventsListener = new EventsListener();
     }
 
@@ -87,7 +79,7 @@ public class LegacyRedisStorageEngine extends AbstractEventStorageEngine {
         log.info("Starting legacy redis storage.");
 
         // check si on a l'index des events
-        eventsIndex.rebuildIfNecessary();
+        eventsIndexer.rebuildIfNecessary();
 
         // se met à l'écoute des nouveaux events.
         this.eventsListener.start();
@@ -270,134 +262,6 @@ public class LegacyRedisStorageEngine extends AbstractEventStorageEngine {
         }
     }
 
-
-    /**
-     * index les events
-     */
-    private class EventsIndex {
-
-        static final String A_EVENTS_INDEX_LIST = "a_events_index_list";
-        static final String A_EVENTS_INDEX_SET  = "a_events_index_set";
-        final DefaultRedisScript<Long> indexEventsScript = new DefaultRedisScript<>();
-
-        EventsIndex() {
-            indexEventsScript.setScriptSource(new ResourceScriptSource(new ClassPathResource("redis/index_events.lua")));
-            indexEventsScript.setResultType(Long.class);
-        }
-
-        private List<String> retrieveAllKeys(StopWatch watch) {
-            watch.start("retrieve all keys ");
-            List<String> keys = template.keys("*").stream().filter(this::isAEventSourcedKey).collect(Collectors.toList());
-            watch.stop();
-            return keys;
-        }
-
-        private void restoreSaveConfiguration(String currentSave) {
-            template.getConnectionFactory().getConnection().setConfig("save", currentSave);
-        }
-
-        private void disableRdb(StopWatch watch) {
-            watch.start("Disable redis 'save' processing");
-            template.getConnectionFactory().getConnection().setConfig("save", "");
-            watch.stop();
-        }
-
-        private String getCurrentSaveConfiguration(StopWatch watch) {
-            watch.start("Saving redis current 'save' configuration");
-            List<String> saveList = template.getConnectionFactory().getConnection().getConfig("save");
-            String currentSave = "";
-            if (!saveList.isEmpty()) {
-                currentSave = saveList.get(1);
-            }
-            log.info("Saving 'save' configuration : {}", currentSave);
-            watch.stop();
-            return currentSave;
-        }
-
-        private boolean isAEventSourcedKey(Object key) {
-            return key.toString().startsWith("module-") || key.toString().startsWith("platform-") || key.toString().startsWith("template_package-");
-        }
-
-        void rebuildIfNecessary() {
-
-//            if (template.hasKey(A_EVENTS_INDEX_LIST) && template.hasKey(A_EVENTS_INDEX_SET)) {
-//                log.info("no need to rebuild index ...");
-//                return;
-//            }
-            // rebuild l'index ici.
-            StopWatch watch = new StopWatch();
-
-            List<String> keys = retrieveAllKeys(watch);
-
-            // desactivation du système de persistence, parce qu'on va ajouter beaucoup de clé !!
-            String currentSave = getCurrentSaveConfiguration(watch);
-
-            // desactive rdb
-            disableRdb(watch);
-
-            watch.start("indexing aggregates.");
-            MultiThreadedIndexer indexer = new MultiThreadedIndexer();
-            indexer.process(keys);
-            watch.stop();
-
-            // restore rdb
-            restoreSaveConfiguration(currentSave);
-
-            log.info("{}", watch.prettyPrint());
-            log.info("Processed: keys: {}, noop: {}, ok: {}", indexer.keyCount, indexer.noopCount, indexer.okCount);
-            log.info("done. total time: {} ms", watch.getTotalTimeMillis());
-        }
-
-        class MultiThreadedIndexer {
-
-            AtomicInteger keyCount = new AtomicInteger();
-            AtomicInteger okCount = new AtomicInteger();
-            AtomicInteger koCount = new AtomicInteger();
-            AtomicInteger noopCount = new AtomicInteger();
-
-            void process(List<String> keys) {
-
-                int processorsCount = 50;
-                List<List<String>> batches = Lists.partition(keys, keys.size() / processorsCount);
-
-                ExecutorService threadPool = Executors.newFixedThreadPool(batches.size() + 2);
-
-                List<Callable<Object>> callables = batches.stream().map(batch -> Executors.callable(() -> doWithBatchOfKeys(batch))).collect(Collectors.toList());
-
-                // thread de contrôle.
-                ScheduledExecutorService controlService = Executors.newSingleThreadScheduledExecutor();
-                controlService.scheduleAtFixedRate(() -> log.info("current key count {}, ok={}, ko={}, noop={}",
-                        keyCount.get(), okCount.get(), koCount.get(), noopCount.get()), 1, 1, TimeUnit.SECONDS);
-
-                try {
-                    threadPool.invokeAll(callables);
-                } catch (InterruptedException e) {
-                    log.error(e.getMessage(), e);
-                    Thread.currentThread().interrupt();
-                }
-                controlService.shutdownNow();
-            }
-
-            void doWithBatchOfKeys(List<String> batch) {
-
-                for (String key : batch) {
-                    keyCount.incrementAndGet();
-                    try {
-                        long result = template.execute(indexEventsScript, Collections.singletonList(key));
-                        if (result > 0) {
-                            okCount.incrementAndGet();
-                        } else {
-                            noopCount.incrementAndGet();
-                        }
-                    } catch (Exception e) {
-                        log.error(e.getMessage(), e);
-                        koCount.incrementAndGet();
-                    }
-                }
-            }
-        }
-
-    }
 
     /**
      * écoute les nouveaux events. utilise le système de pub/sub de Redis
